Fold id-only bot message ops into existing controllers
