--- app/controllers/api/bots/messages_by_id/boosts_controller.rb
@@ -1,33 +1,0 @@
-class API::Bots::MessagesById::BoostsController < API::Bots::BaseController
-  include NotifyBots
-
-  before_action :set_message_with_access_check
-
-  def create
-    content = request.body.tap(&:rewind).read.force_encoding("UTF-8").strip
-    @boost = @message.boosts.create!(content: content, booster: Current.user)
-
-    notify_bots(@boost, :created)
-
-    render json: { id: @boost.id, content: @boost.content }, status: :created
-  rescue ActiveRecord::RecordNotUnique, ActiveRecord::RecordInvalid
-    render json: { error: "Boost already exists or is invalid", code: "validation_failed" }, status: :unprocessable_entity
-  end
-
-  def destroy
-    @boost = @message.boosts.where(booster: Current.user).find(params[:id])
-
-    @boost.destroy!
-    notify_bots(@boost, :deleted)
-
-    head :no_content
-  end
-
-  private
-    def set_message_with_access_check
-      @message = Message.active
-                        .where(room_id: Current.user.rooms.select(:id))
-                        .find(params[:message_id])
-      @room = @message.room
-    end
-end
